Added Some Commands and Repair Script
Added embed: It will send an embed of whatever message you'd like to your current channel.

Added imitate: It will imitate and user with any message in the chat you are currently in.

Added sportsbetting module: Commands are /add, /showslip, /showoff, and /win.

Added wpost: Adds a welcome post to the welcome channel also shows who the owner of the server is.

Added voice-channel-repair.bat: Run this if you are on windows and your bot keeps joining and leaving the voice channel when using any of the voice commands. I will soon embed this function into the bot itself. This is a temporary fix.